+++ components/script/dom/globalscope.rs
@@ -26,6 +26,7 @@ use crate::dom::window::Window;
 use crate::dom::workerglobalscope::WorkerGlobalScope;
 use crate::dom::workletglobalscope::WorkletGlobalScope;
 use crate::microtask::{Microtask, MicrotaskQueue};
+use crate::script_ipc_router::{ScriptIpcRouter, IpcCallback};
 use crate::script_runtime::{CommonScriptMsg, JSContext as SafeJSContext, ScriptChan, ScriptPort};
 use crate::script_thread::{MainThreadScriptChan, ScriptThread};
 use crate::task::TaskCanceller;
@@ -38,11 +39,9 @@ use crate::task_source::websocket::WebsocketTaskSource;
 use crate::task_source::TaskSourceName;
 use crate::timers::{IsInterval, OneshotTimerCallback, OneshotTimerHandle};
 use crate::timers::{OneshotTimers, TimerCallback};
-use crossbeam_channel::{unbounded, Sender};
 use devtools_traits::{ScriptToDevtoolsControlMsg, WorkerId};
 use dom_struct::dom_struct;
-use ipc_channel::ipc::{self, IpcSender};
-use ipc_channel::router::ROUTER;
+use ipc_channel::ipc::IpcSender;
 use js::glue::{IsWrapper, UnwrapObjectDynamic};
 use js::jsapi::JSObject;
 use js::jsapi::{CurrentGlobalOrNull, GetNonCCWObjectGlobal};
@@ -53,7 +52,7 @@ use js::rust::wrappers::EvaluateUtf8;
 use js::rust::{get_object_class, CompileOptionsWrapper, ParentRuntime, Runtime};
 use js::rust::{HandleValue, MutableHandleValue};
 use js::{JSCLASS_IS_DOMJSCLASS, JSCLASS_IS_GLOBAL};
-use msg::constellation_msg::{IpcCallbackId, IpcCallbackMsg, IpcHandle, PipelineId};
+use msg::constellation_msg::{IpcHandle, PipelineId};
 use net_traits::image_cache::ImageCache;
 use net_traits::{CoreResourceThread, IpcSend, ResourceThreads};
 use profile_traits::{mem as profile_mem, time as profile_time};
@@ -70,64 +69,6 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use time::{get_time, Timespec};
 
-pub type IpcCallback = Box<dyn FnMut(Vec<u8>) + Send>;
-
-pub struct IpcScriptRouter {
-    ipc_sender: IpcSender<IpcCallbackMsg>,
-    sender: Sender<(IpcCallbackId, IpcCallback)>,
-}
-
-impl IpcScriptRouter {
-    fn new() -> Self {
-        let (callback_sender, callback_receiver) =
-            ipc::channel().expect("IpcScriptRouter ipc chan");
-        let (sender, receiver) = unbounded();
-        let ipc_script_router = IpcScriptRouter {
-            sender: sender,
-            ipc_sender: callback_sender,
-        };
-        let mut callbacks: HashMap<IpcCallbackId, IpcCallback> = HashMap::new();
-        ROUTER.add_route(
-            callback_receiver.to_opaque(),
-            Box::new(move |message| {
-                match message.to().expect("IpcScriptRouter handle incoming msg") {
-                    IpcCallbackMsg::AddCallback => {
-                        let (id, callback) =
-                            receiver.recv().expect("Callback message to be received");
-                        callbacks.insert(id, callback);
-                    },
-                    IpcCallbackMsg::DropCallback(id) => {
-                        let _ = callbacks
-                            .remove(&id)
-                            .expect("Callbackt to be removed to exists");
-                    },
-                    IpcCallbackMsg::Callback(id, data) => {
-                        let callback = callbacks
-                            .get_mut(&id)
-                            .expect("Callback to be called to exists");
-                        callback(data);
-                    },
-                }
-            }),
-        );
-        ipc_script_router
-    }
-
-    fn add_callback(&self, callback: IpcCallback) -> IpcHandle {
-        let callback_id = IpcCallbackId::new();
-        if let Ok(_) = self.sender.send((callback_id.clone(), callback)) {
-            self.ipc_sender
-                .send(IpcCallbackMsg::AddCallback)
-                .expect("The script ipc router to be available");
-            return IpcHandle {
-                callback_id,
-                sender: self.ipc_sender.clone(),
-            };
-        }
-        unreachable!("Adding an ipc callback failed");
-    }
-}
-
 #[derive(JSTraceable)]
 pub struct AutoCloseWorker(Arc<AtomicBool>);
 
@@ -137,7 +78,7 @@ impl Drop for AutoCloseWorker {
     }
 }
 
-unsafe_no_jsmanaged_fields!(IpcScriptRouter);
+unsafe_no_jsmanaged_fields!(ScriptIpcRouter);
 
 #[dom_struct]
 pub struct GlobalScope {
@@ -147,7 +88,7 @@ pub struct GlobalScope {
 
     /// A shared-router for script to handle ipc.
     #[ignore_malloc_size_of = "channels are hard"]
-    ipc_script_router: IpcScriptRouter,
+    ipc_script_router: ScriptIpcRouter,
 
     /// Pipeline id associated with this global.
     pipeline_id: PipelineId,
@@ -248,7 +189,7 @@ impl GlobalScope {
         user_agent: Cow<'static, str>,
     ) -> Self {
         Self {
-            ipc_script_router: IpcScriptRouter::new(),
+            ipc_script_router: ScriptIpcRouter::new(),
             eventtarget: EventTarget::new_inherited(),
             crypto: Default::default(),
             next_worker_id: Cell::new(WorkerId(0)),
